backend/pyproject.toml
[project]
name = "wrong-by-default"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp==3.11.0",
    "fastapi==0.115.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "python-dotenv==1.0.1",
    "uvicorn[standard]==0.32.0",
]
